+++ musicdl/modules/sources/migu.py
@@ -49,7 +49,9 @@ class MiguMusicClient(BaseMusicClient):
         boost_result = dict(download_result=download_result, download_url=download_url, file_size=file_size, download_url_status=download_url_status, ext=ext)
         return boost_result
     '''_constructsearchurls'''
-    def _constructsearchurls(self, keyword: str, rule: dict = {}, request_overrides: dict = {}):
+    def _constructsearchurls(self, keyword: str, rule: dict = None, request_overrides: dict = None):
+        # init
+        rule, request_overrides = rule or {}, request_overrides or {}
         # search rules
         default_rule = {
             "text": keyword, 'pageNo': 1, 'pageSize': 10
@@ -67,7 +69,9 @@ class MiguMusicClient(BaseMusicClient):
         return search_urls
     '''_search'''
     @usesearchheaderscookies
-    def _search(self, keyword: str = '', search_url: str = '', request_overrides: dict = {}, song_infos: list = [], progress: Progress = None, progress_id: int = 0):
+    def _search(self, keyword: str = '', search_url: str = '', request_overrides: dict = None, song_infos: list = [], progress: Progress = None, progress_id: int = 0):
+        # init
+        request_overrides = request_overrides or {}
         # _safefetchfilesize
         def _safefetchfilesize(meta: dict):
             file_size = meta.get('asize') or meta.get('isize') or meta.get('size') or '0'