+++ musicdl/modules/sources/ximalaya.py
@@ -47,7 +47,9 @@ class XimalayaMusicClient(BaseMusicClient):
         plaintext = re.sub(r"[^\x20-\x7E]", "", plaintext.decode("utf-8"))
         return plaintext
     '''_constructsearchurls'''
-    def _constructsearchurls(self, keyword: str, rule: dict = {}, request_overrides: dict = {}):
+    def _constructsearchurls(self, keyword: str, rule: dict = None, request_overrides: dict = None):
+        # init
+        rule, request_overrides = rule or {}, request_overrides or {}
         # search rules
         default_rule = {'msg': keyword, 'n': '', 'num': self.search_size_per_source, 'type': 'json'}
         default_rule.update(rule)
@@ -63,7 +65,9 @@ class XimalayaMusicClient(BaseMusicClient):
         return search_urls
     '''_search'''
     @usesearchheaderscookies
-    def _search(self, keyword: str = '', search_url: str = '', request_overrides: dict = {}, song_infos: list = [], progress: Progress = None, progress_id: int = 0):
+    def _search(self, keyword: str = '', search_url: str = '', request_overrides: dict = None, song_infos: list = [], progress: Progress = None, progress_id: int = 0):
+        # init
+        request_overrides = request_overrides or {}
         # successful
         try:
             # --search results